agent/jd_skill_processor_agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.34.0
requests>=2.31.0
bedrock-agentcore==0.1.7
httpx>=0.27.0
botocore>=1.34.0


